# src/main/java/com/evbooking/dto/UserResponse.java
package com.evbooking.dto;

import com.evbooking.model.User;

public record UserResponse(Long id, String username, String role) {

    public static UserResponse from(User user) {
        return new UserResponse(user.getId(), user.getUsername(), user.getRole().name());
    }
}