# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: EV Charging Station Booking API
  version: 1.0.0
  description: |
    REST API for the EV Charging Station Booking System.

    **Authentication** is session-based. Call `POST /api/auth/login` to receive a
    `JSESSIONID` cookie; include it in every subsequent request.

    **Roles**
    | Role   | Can do                                                         |
    |--------|----------------------------------------------------------------|
    | DRIVER | Register, log in, book connectors, manage own bookings.        |
    | ADMIN  | Everything a driver can do, plus manage stations, connectors, and all bookings. |
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://ev-charging-booking.herokuapp.com
    description: Production (Heroku)
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Login
      description: |
        Authenticates the user and creates an HTTP session.
        The response sets a `JSESSIONID` cookie which must be included in all
        subsequent requests to protected endpoints.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: alice
              password: secret123
      responses:
        '200':
          description: Login successful — returns the authenticated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid username or password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InlineError'
              example:
                error: "Invalid username or password."
  /api/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Logout
      description: Invalidates the current session. Subsequent requests with the same cookie will be treated as unauthenticated.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Session invalidated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InlineMessage'
              example:
                message: "Logged out successfully."
  /api/auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Get current user
      description: Returns the profile of the currently authenticated user.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: The authenticated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: No active session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InlineError'
              example:
                error: "Not authenticated."
  /api/auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Register
      description: |
        Creates a new user account with the **DRIVER** role.
        Self-registration never grants the ADMIN role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              username: newdriver
              password: pass1234
      responses:
        '201':
          description: Account created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  # ── Stations ──────────────────────────────────────────────────────────────
...
  /api/bookings:
    get:
      tags: [Bookings]
      operationId: listMyBookings
      summary: List my bookings
      description: Returns all bookings owned by the currently authenticated driver.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Array of the driver's bookings.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookingResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Bookings]
      operationId: createBooking
      summary: Create a booking
      description: |
        Books a connector time slot for the authenticated driver.
        Returns **409 Conflict** if the slot overlaps with an existing active booking.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRequest'
            example:
              connectorId: 3
              bookingDate: "2025-06-20"
              startTime: "10:00:00"
              endTime: "11:30:00"
      responses:
        '201':
          description: Booking created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/bookings/{id}:
    get:
      tags: [Bookings]
      operationId: getBooking
      summary: Get a booking
      description: Returns a single booking by ID. Drivers can only retrieve their own bookings.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: The booking.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Bookings]
      operationId: rescheduleBooking
      summary: Reschedule a booking
      description: |
        Moves an existing booking to a new date and time window on the same connector.
        Returns **409 Conflict** if the new slot overlaps with another active booking.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RescheduleRequest'
            example:
              bookingDate: "2025-06-21"
              startTime: "14:00:00"
              endTime: "15:00:00"
      responses:
        '200':
          description: Booking rescheduled — returns the updated booking.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Bookings]
      operationId: cancelBooking
      summary: Cancel a booking
      description: |
        Cancels a booking. Drivers may only cancel their own future bookings.
        Admins may cancel any booking regardless of owner or status.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: "Booking cancelled — returns the updated booking with status: CANCELLED."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Admin ─────────────────────────────────────────────────────────────────
  /api/admin/users:
    get:
      tags: [Admin]
      operationId: adminListUsers
      summary: List all users
      description: Returns every registered user in the system. Requires **ADMIN** role.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Array of all users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/bookings:
    get:
      tags: [Admin]
      operationId: adminListBookings
      summary: List all bookings
      description: Returns every booking in the system regardless of owner. Requires **ADMIN** role.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Array of all bookings.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookingResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/bookings/{id}:
    delete:
      tags: [Admin]
      operationId: adminCancelBooking
      summary: Cancel any booking
      description: Allows an admin to cancel any booking regardless of its owner or status. Requires **ADMIN** role.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: "Booking cancelled — returns the updated booking with status: CANCELLED."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

# ─────────────────────────────────────────────────────────────────────────────
# Components
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Session cookie obtained after a successful `POST /api/auth/login`.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: Resource ID.
      schema:
        type: integer
        format: int64
        example: 1
  responses:
    BadRequest:
      description: Validation failed or request is malformed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 400
            error: "Bad Request"
            message: "must not be blank"
            timestamp: "2025-06-15T09:30:00"
    Unauthorized:
      description: No active session — authentication required.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 401
            error: "Unauthorized"
            message: "Not authenticated."
            timestamp: "2025-06-15T09:30:00"
    Forbidden:
      description: Authenticated but lacks the required role.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 403
            error: "Forbidden"
            message: "You do not have permission to perform this action."
            timestamp: "2025-06-15T09:30:00"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 404
            error: "Not Found"
            message: "Station not found with id: 99"
            timestamp: "2025-06-15T09:30:00"
    Conflict:
      description: Resource conflict — e.g. booking time overlap or duplicate username.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 409
            error: "Conflict"
            message: "Connector 3 is already booked between 10:00 and 11:30 on 2025-06-20."
            timestamp: "2025-06-15T09:30:00"
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: alice
        password:
          type: string
          format: password
          minLength: 4
          maxLength: 72
          example: secret123
    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: newdriver
        password:
          type: string
          format: password
          minLength: 4
          maxLength: 72
          example: pass1234
...
    BookingRequest:
      type: object
      required: [connectorId, bookingDate, startTime, endTime]
      properties:
        connectorId:
          type: integer
          format: int64
          example: 3
        bookingDate:
          type: string
          format: date
          example: "2025-06-20"
        startTime:
          type: string
          description: ISO-8601 local time (`HH:mm:ss`).
          example: "10:00:00"
        endTime:
          type: string
          description: ISO-8601 local time (`HH:mm:ss`). Must be after `startTime`.
          example: "11:30:00"
    RescheduleRequest:
      type: object
      required: [bookingDate, startTime, endTime]
      properties:
        bookingDate:
          type: string
          format: date
          example: "2025-06-21"
        startTime:
          type: string
          description: ISO-8601 local time (`HH:mm:ss`).
          example: "14:00:00"
        endTime:
          type: string
          description: ISO-8601 local time (`HH:mm:ss`). Must be after `startTime`.
          example: "15:00:00"

    # ── Response bodies ──────────────────────────────────────────────────────
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: alice
        role:
          type: string
          enum: [DRIVER, ADMIN]
          example: DRIVER
...
    BookingResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 42
        userId:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: alice
        connectorId:
          type: integer
          format: int64
          example: 3
        connectorType:
          type: string
          example: "CCS2"
        stationId:
          type: integer
          format: int64
          example: 1
        stationName:
          type: string
          example: "City Centre Hub"
        bookingDate:
          type: string
          format: date
          example: "2025-06-20"
        startTime:
          type: string
          example: "10:00:00"
        endTime:
          type: string
          example: "11:30:00"
        status:
          type: string
          enum: [ACTIVE, CANCELLED, COMPLETED]
          example: ACTIVE
        createdAt:
          type: string
          format: date-time
          example: "2025-06-15T09:30:00"
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 404
        error:
          type: string
          example: "Not Found"
        message:
          type: string
          example: "Station not found with id: 99"
        timestamp:
          type: string
          format: date-time
          example: "2025-06-15T09:30:00"
    InlineError:
      type: object
      description: Simple error wrapper used by a few auth endpoints.
      properties:
        error:
          type: string
          example: "Invalid username or password."
    InlineMessage:
      type: object
      description: Simple success message wrapper.
      properties:
        message:
          type: string
          example: "Logged out successfully."